Amplifier based on TDA2003
This is the first circuit I've done.
This circuit is based on the test circuit that comes in the datasheet of the TDA2003 but which have removed some components. Nevertheless, has a quality of sound acceptable.
Ideal for carrying music festivals fed by a small battery.
My electronics teacher Txitxi, showed me this circuit, but must also be in the www.